update zip to reflect updated repo files as of June 23 2025
The zip that is there now (cf_training_exercices_2021.zip) hasn't been updated in 3 yrs, and so it doesn't reflect the many code (and folder) updates which have been done in the repo since then. There are many changes and some additions.

To be clear, I kept the zip having a top-level folder called "cf_specialist_exercises", even though the repo doesn't use that folder name. I also did correct that typo in the folder name, just like someone had modified the code in the repo to correct its subfolder "exercices", and references to it. :-)